A few necessary changes...

+++ Platforms/MaxelerFPGA/maxeler_code/src/mc_solver_maxeler/CombinedTauswortheRNG.java
@@ -15,7 +15,7 @@ public class CombinedTauswortheRNG extends KernelLib{
   protected Kernel kernel;
   public DFEVar s1,s2,s3,offset,count;
 
-  public CombinedTauswortheRNG(Kernel kernel,int period,DFEVar s1,DFEVar s2,DFEVar s3,DFEVar offset){    
+  public CombinedTauswortheRNG(Kernel kernel,long period,DFEVar s1,DFEVar s2,DFEVar s3,DFEVar offset){    
       super(kernel);
       
       this.kernel = kernel;
